workers/requirements.txt
# Workers Python dependencies
celery==5.4.0
redis==5.0.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
boto3==1.35.0
python-dotenv==1.0.1

# AI & ML
openai==1.50.0
anthropic==0.34.0
openai-whisper==20231117

# Audio analysis
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4

# Video analysis
scenedetect[opencv]==0.6.4
opencv-python-headless==4.10.0.84
moviepy==1.0.3
Pillow==10.4.0
pytesseract==0.3.13
scikit-image==0.24.0
scikit-learn==1.5.0

# Utilities
httpx==0.27.0
pydantic==2.9.0
